Crate matrix_sdk_common

Common types that are used in the different matrix-sdk layers.

NOTE: These are common dependencies for the different matrix-sdk layers, you’re probably interested in the main matrix-sdk crate.

Re-exports

pub use instant;

Modules

deserialized_responses

executor

Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

locks

timeout

Traits

AsyncTraitDeps

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

SendOutsideWasm

Alias for Send on non-wasm, empty trait (implemented by everything) on wasm.

SyncOutsideWasm

Alias for Sync on non-wasm, empty trait (implemented by everything) on wasm.