Module matrix_sdk_common::debug

Helpers for creating std::fmt::Debug implementations.

Structs

DebugRawEvent

A wrapper around Raw that implements Debug in a way that only prints the event ID and event type.

DebugRawEventNoId

A wrapper around Raw that implements Debug in a way that only prints the event type.

Traits

DebugStructExt