Module matrix_sdk_common::deserialized_responses

Structs

AccountData
Data that the user has attached to either the account or a specific room.

AmbiguityChange
A change in ambiguity of room members that an m.room.member event triggers.

AmbiguityChanges
Collection of ambiguioty changes that room member events trigger.

Ephemeral
The ephemeral events in the room that aren’t recorded in the timeline or state of the room. e.g. typing.

InviteState
The state of a room that the user has been invited to.

InvitedRoom
Updates to the rooms that the user has been invited to.

JoinedRoom
Updates to joined rooms.

LeftRoom

MemberEvent

MembersResponse
A deserialized response for the rooms members API call.

Presence
Updates to the presence status of other users.

Rooms

State
State events in the room.

StrippedMemberEvent

SyncResponse

Timeline
Events in the room.

ToDevice
Messages sent dirrectly between devices.

UnreadNotificationsCount
Counts of unread notifications for a room.