Enum matrix_sdk_common::deserialized_responses::DeviceLinkProblem

pub enum DeviceLinkProblem {
    MissingDevice,
    InsecureSource,
}

The sub-enum containing detailed information on why we were not able to link a message back to a device.

Variants

MissingDevice

The device is missing, either because it was deleted, or you haven’t yet downoaled it or the server is erroneously omitting it (federation lag).

InsecureSource

The key was obtained from an insecure source: imported from a file, obtained from a legacy (asymmetric) backup, unsafe key forward, etc.