Crate matrix_sdk_common

Common types that are used in the different matrix-sdk layers.

NOTE: These are common dependencies for the different matrix-sdk layers, you’re probably interested in the main matrix-sdk crate.

Re-exports

pub use instant;

pub use ruma;

pub use store_locks::LEASE_DURATION_MS;

Modules

debug

Helpers for creating std::fmt::Debug implementations.

deserialized_responses

executor

Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

failures_cache

A TTL cache which can be used to time out repeated operations that might experience intermittent failures.

ring_buffer

store_locks

Collection of small helpers that implement store-based locks.

timeout

tracing_timer

Macros

boxed_into_future

timer

Macro to create a RAII timer that will log a tracing event once it’s dropped.

Traits

AsyncTraitDeps

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

SendOutsideWasm

Alias for Send on non-wasm, empty trait (implemented by everything) on wasm.

SyncOutsideWasm

Alias for Sync on non-wasm, empty trait (implemented by everything) on wasm.