Module matrix_sdk_common::deserialized_responses

Structs

AmbiguityChange
A change in ambiguity of room members that an m.room.member event triggers.

AmbiguityChanges
Collection of ambiguioty changes that room member events trigger.

EncryptionInfo
Struct containing information on how an event was decrypted.

JoinedRoom
Updates to joined rooms.

LeftRoom

MemberEvent

MembersResponse
A deserialized response for the rooms members API call.

Rooms

StrippedMemberEvent

SyncResponse

SyncRoomEvent
A customized version of a room event coming from a sync that holds optional encryption info.

Timeline
Events in the room.

UnreadNotificationsCount
Counts of unread notifications for a room.

Enums

AlgorithmInfo
The algorithm specific information of a decrypted event.

VerificationState
The verification state of the device that sent an event to us.