Trait matrix_sdk_common::Outgoing

pub trait Outgoing {
    type Incoming;
}

A type that can be sent to another party that understands the matrix protocol.

If any of the fields of Self don’t implement serde’s Deserialize, you can derive this trait to generate a corresponding ‘Incoming’ type that supports deserialization. This is useful for things like ruma_events’ EventResult type. For more details, see the derive macro’s documentation.

Associated Types

type Incoming

The ‘Incoming’ variant of Self.

Loading content...